Revert uint8_t deltas.
Some deltas for 347# are larger :/
Tried a variety of +255 hacks again and it's all ugly.
There is a memory leak that needs to be fixed but otherwise I don't
think memory size is a big issue.

+++ combined_sieve_small.h
@@ -36,7 +36,7 @@ class SieveOutput {
         // m_increment, unknown count
         vector<std::tuple<int16_t, int16_t>> m_inc;
 
-        vector<vector<uint8_t>> unknowns;
+        vector<vector<uint16_t>> unknowns;
 };
 
 std::unique_ptr<SieveOutput> prime_gap_parallel(const struct Config& config);
